+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Change Log
 - Support Revenue by product[`#28 <https://github.com/defeat-beta/defeatbeta-api/issues/28>`_]
 - Support Quarterly/Annual Operating Margin %[`#18 <https://github.com/defeat-beta/defeatbeta-api/issues/18>`_]
 - Support Quarterly/Annual Net Margin %[`#20 <https://github.com/defeat-beta/defeatbeta-api/issues/20>`_]
+- Support Quarterly/Annual EBITDA Margin %[`#19 <https://github.com/defeat-beta/defeatbeta-api/issues/19>`_]
 
 0.0.9
 -------